Fix update progress type (#3767)
* fix(jenkins): update build stages logic (#3760)

* feat(dora): update logic for change lead time (#3742)

* feat(dora): update logic for change lead time

closes #3516

* feat(dora): update e2e according to review

* fix(zentao): update progress typoe

+++ plugins/zentao/models/archived/product.go
@@ -52,7 +52,7 @@ type ZentaoProduct struct {
 	Executions     int        `json:"executions"`
 	Bugs           int        `json:"bugs"`
 	Docs           int        `json:"docs"`
-	Progress       int        `json:"progress"`
+	Progress       float64    `json:"progress"`
 	CaseReview     bool       `json:"caseReview"`
 	archived.NoPKModel
 }
